Implements intersecting_products for u32
Changes generated samples to this type to allow some benchmarks. Early results
suggest calling out to rust is slower than staying in python which makes sense
considering that relatively few steps of computation need to happen for each output
and in the rust implementation each output and input element must be converted.

The tests also expose a bug in either the python or rust implementation since they do
not always agree.